prepare 2.2.0 release (#40)
## [2.2.0] - 2024-05-01
### Added:
- Added an optional timeout to the `waitForInitialization` method. When
a timeout is specified the returned promise will be rejected after the
timeout elapses if the client has not finished initializing within that
time. When no timeout is specified the returned promise will not be
resolved or rejected until the initialization either completes or fails.

### Changed:
- The track method now validates that the provided metricValue is a
number. If a metric value is provided, and it is not a number, then a
warning will be logged.

### Fixed:
- Fixed the documentation for `evaluationReasons` for the `identify`
method.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Louis Chan <91093020+louis-launchdarkly@users.noreply.github.com>
Co-authored-by: LaunchDarklyReleaseBot <launchdarklyreleasebot@launchdarkly.com>
Co-authored-by: Matthew M. Keeler <mkeeler@launchdarkly.com>
Co-authored-by: Ryan Lamb <4955475+kinyoklion@users.noreply.github.com>
Co-authored-by: Casey Waldren <cwaldren@launchdarkly.com>